+++ Patches/CustomCameraControllerPatch.cs
@@ -5,39 +5,27 @@ namespace BetterSMT.Patches;
 
 [HarmonyPatch(typeof(CustomCameraController), "LateUpdate")]
 public class LateUpdateRaycastPatch {
-
-
     private static bool isThirdPersonEnabled = false;
 
     private static void Postfix(CustomCameraController __instance) {
-        if (!FsmVariables.GlobalVariables.GetFsmBool("InChat").Value == false) {
+        if (FsmVariables.GlobalVariables.GetFsmBool("InChat").Value) {
             return;
-        } else {
-            if (BetterSMT.ThirdPersonToggle.Value == true) {
-                if (__instance == null || __instance.inVehicle || __instance.isInCameraEvent) {
-                    return;
-                }
-
-                if (BetterSMT.ThirdPersonHotkey.Value.IsDown()) {
-                    isThirdPersonEnabled = !isThirdPersonEnabled;
-                }
-
-                if (isThirdPersonEnabled) {
-                    __instance.inEmoteEvent = true;
-                    if (__instance.thirdPersonFollow != null) {
-                        __instance.thirdPersonFollow.CameraDistance = 2f;
-                        __instance.thirdPersonFollow.CameraCollisionFilter = __instance.thirdPersonDefaultLayerMask;
-                    }
-                    __instance.ShowCharacter(true);
-                } else {
-                    __instance.inEmoteEvent = false;
-                    if (__instance.thirdPersonFollow != null) {
-                        __instance.thirdPersonFollow.CameraDistance = 0f;
-                        __instance.thirdPersonFollow.CameraCollisionFilter = 0;
-                    }
-                    __instance.ShowCharacter(false);
-                }
-            }
         }
+
+        if (!BetterSMT.ThirdPersonToggle.Value || __instance == null || __instance.inVehicle || __instance.isInCameraEvent) {
+            return;
+        }
+
+        if (BetterSMT.ThirdPersonHotkey.Value.IsDown()) {
+            isThirdPersonEnabled = !isThirdPersonEnabled;
+        }
+
+        __instance.inEmoteEvent = isThirdPersonEnabled;
+        if (__instance.thirdPersonFollow != null) {
+            __instance.thirdPersonFollow.CameraDistance = isThirdPersonEnabled ? 2f : 0f;
+            __instance.thirdPersonFollow.CameraCollisionFilter = isThirdPersonEnabled ? __instance.thirdPersonDefaultLayerMask : 0;
+        }
+
+        __instance.ShowCharacter(isThirdPersonEnabled);
     }
 }