+++ Patches/ProductListingPatch.cs
@@ -1,6 +1,5 @@
 ﻿using HarmonyLib;
 using System.Collections.Generic;
-using System.Linq;
 using UnityEngine;
 
 namespace BetterSMT.Patches;